Fix Atom feed links
The URLs in http://blog.jwilm.io/feed.xml point an example domain. I'm not familiar with Ruby, nor sure this is the correct branch, but hopefully this will fix it?

Normally I would just e-mail you, but I stumbled across this when trying to find your e-mail address, and, well, here we are. Thanks for the interesting blog!